+++ Cargo.toml
@@ -3,15 +3,7 @@ name = "impact"
 version = "0.1.1"
 authors = ["deciduously <github@deciduously.com>"]
 
-[[bin]]
-name = "impact"
-path = "src/main.rs"
-
-[lib]
-name = "impact"
-path = "src/lib.rs"
-
 [dependencies]
 
 stdweb = "0.4"
-yew = { git = "https://github.com/DenisKolodin/yew" }
+yew = "0.7"
