+++ package.json
@@ -7,34 +7,31 @@
   "author": "deciduously <github@deciduously.com>",
   "license": "MIT",
   "scripts": {
-    "clean": "cargo clean",
-    "clean:deploy": "rm -rf release/",
     "build:rs": "cargo web deploy --release",
     "build:js": "rollup -c",
-    "build:scss": "node-sass --include-path scss scss/impact.scss css/impact.css",
-    "build:css": "postcss --use autoprefixer -o static/impact.css css/impact.css",
     "build:copy": "cp target/deploy/impact.css release/ && cp target/deploy/impact.wasm release/ && cp target/deploy/index.html release/ && cp target/deploy/favicon.ico release/",
     "build": "run-s clean:deploy build:rs build:js build:copy",
-    "serve": "serve -p 8080 release/",
+    "clean": "cargo clean",
+    "clean:deploy": "rm -rf release/",
+    "dev": "run-s style watch:rs",
     "prod": "run-s build serve",
-    "watch:rs": "cargo web start --release",
-    "watch:scss": "nodemon -e scss -x \"npm run build:scss\"",
-    "watch:css": "nodemon -e css -x \"npm run build:css\"",
-    "watch:style": "run-s watch:scss watch:css",
-    "watch": "run-p watch:rs watch:scss watch:style",
-    "start": "run-s watch",
-    "test": "echo \"Error: no tests!\" && exit 1"
+    "serve": "serve -p 8080 release/",
+    "style": "postcss scss/*.scss > css/impact.css",
+    "test": "echo \"Error: no tests!\" && exit 1",
+    "watch:rs": "cargo web start --release"
   },
   "devDependencies": {
     "@babel/core": "^7.8.4",
     "@babel/preset-env": "^7.8.4",
+    "@fullhuman/postcss-purgecss": "^2.0.6",
     "@rollup/plugin-wasm": "^3.0.0",
     "autoprefixer": "^9.7.4",
+    "cssnano": "^4.1.10",
     "node-sass": "^4.13.1",
-    "nodemon": "^2.0.2",
     "npm-run-all": "^4.1.5",
     "postcss": "^7.0.27",
     "postcss-cli": "^7.1.0",
+    "postcss-scss": "^2.0.0",
     "rollup": "^1.31.1",
     "rollup-plugin-babel": "^4.3.3",
     "rollup-plugin-postcss": "^2.0.6",
